+++ lib/phauxth/authenticate_token.ex
@@ -2,8 +2,9 @@ defmodule Phauxth.AuthenticateToken do
   @moduledoc """
   Authenticates the user by verifying a Phauxth token.
 
-  You need to define a `get_by(attrs)` function in the user_context
-  module you are using (which is MyApp.Accounts by default).
+  You need to define a `get_by(attrs)` function in the `session_module`
+  module you are using - see the documentation for Phauxth.Config
+  for more information about the `session_module`.
 
   ## Token authentication
 
@@ -11,12 +12,16 @@ defmodule Phauxth.AuthenticateToken do
   Phauxth.Token to check that it is valid. If it is valid, user information
   is retrieved from the database.
 
+  If you want to store the token in a cookie, see the documentation for
+  Phauxth.Authenticate.Token, which has an example of how you can create
+  a custom module to verify tokens stored in cookies.
+
   ## Options
 
   There are two options:
 
-    * `:user_context` - the user context module to be used
-      * the default is MyApp.Accounts
+    * `:session_module` - the sessions module to be used
+      * the default is Phauxth.Config.session_module()
     * `:log_meta` - additional custom metadata for Phauxth.Log
       * this should be a keyword list
 
@@ -30,9 +35,9 @@ defmodule Phauxth.AuthenticateToken do
 
       plug Phauxth.AuthenticateToken
 
-  And if you are using a different user context module:
+  And if you are using a different sessions module:
 
-      plug Phauxth.AuthenticateToken, user_context: MyApp.Sessions
+      plug Phauxth.AuthenticateToken, session_module: MyApp.Sessions
 
   In the example above, you need to have the `get_by/1` function
   defined in MyApp.Sessions.