+++ lib/phauxth/token/phx_token.ex
@@ -2,8 +2,6 @@ if Code.ensure_loaded?(Phoenix) do
   defmodule Phauxth.PhxToken do
     @moduledoc """
     Token implementation using Phoenix Token.
-
-    ADD INSTRUCTIONS FOR SIGNING
     """
 
     @behaviour Phauxth.Token
@@ -11,11 +9,20 @@ if Code.ensure_loaded?(Phoenix) do
     alias Phauxth.Config
     alias Phoenix.Token
 
+    @max_age 14_400
+
+    @impl true
+    def sign(data, opts) do
+      key_source = Keyword.get(opts, :key_source, Config.endpoint())
+      salt = Keyword.get(opts, :token_salt, Config.token_salt())
+      Token.sign(key_source, salt, data, opts)
+    end
+
     @impl true
     def verify(token, opts) do
       key_source = Keyword.get(opts, :key_source, Config.endpoint())
       salt = Keyword.get(opts, :token_salt, Config.token_salt())
-      Token.verify(key_source, salt, token, opts ++ [max_age: 14_400])
+      Token.verify(key_source, salt, token, opts ++ [max_age: @max_age])
     end
   end
 end