+++ test/support/authenticate_helper.exs
@@ -2,9 +2,9 @@ defmodule Phauxth.CustomAuthenticate do
   use Phauxth.Authenticate.Base
 
   @impl true
-  def get_user(conn, user_context) do
+  def get_user(conn, session_module) do
     with id when not is_nil(id) <- get_session(conn, :user_id),
-         do: user_context.get_by(%{"user_id" => id})
+         do: session_module.get_by(%{"user_id" => id})
   end
 end
 
@@ -26,5 +26,15 @@ defmodule Phauxth.CustomCall do
   def call(conn, {opts, log_meta}) do
     meta = log_meta ++ [path: conn.request_path]
     super(conn, {opts, meta})
+  end
+end
+
+defmodule Phauxth.AuthenticateTokenCookie do
+  use Phauxth.Authenticate.Token
+
+  @impl true
+  def get_user(%Plug.Conn{req_cookies: %{"access_token" => token}}, opts) do
+    token_mod = Config.token_module()
+    verify_token(token, token_mod, opts)
   end
 end