+++ test/test_helper.exs
@@ -6,6 +6,7 @@ Code.require_file("support/custom_confirm.exs", __DIR__)
 Code.require_file("support/session_helper.exs", __DIR__)
 Code.require_file("support/user_helper.exs", __DIR__)
 
+Application.put_env(:phauxth, :session_module, Phauxth.TestSessions)
 Application.put_env(:phauxth, :endpoint, PhauxthWeb.Endpoint)
 Application.put_env(:phauxth, :token_module, Phauxth.PhxToken)
 Application.put_env(:phauxth, :token_salt, Phauxth.Config.gen_token_salt())